[#2846] Extracted the template-applying flow into 'AbstractInstallCommand'.
Downloading, collecting answers, processing and copying is one flow shared by every verb that applies a template, so it moves to a base class and 'install' becomes a facade over it. Closing guidance is now suppressed on non-interactive runs so a caller's stdout stays clean; build failure output still prints on every path because it explains a non-zero exit code. The header names the operation, which follows the destination state rather than the command name.
